requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
defusedxml==0.7.1
Django==4.2.5
djangorestframework==3.14.0
feedparser==6.0.10
frozenlist==1.4.0
httpie==3.2.2
idna==3.4
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
numpy==1.26.0
pandas==2.1.0
Pygments==2.16.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.5.3
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.4
yarl==1.9.2
